Move actions to seperate file
